Day11 - Flask(Router_CORS_Utils)/requirements.txt
flask
flask-jwt-extended
passlib
mysql-connector-python
